@@ -1,31 +1,43 @@
 {
   "name": "rss",
   "scripts": {
-    "build": "webpack",
-    "server": "webpack-dev-server --hot --inline --progress --colors"
+    "build": "webpack --progress --colors",
+    "start": "webpack-dev-server --open",
+    "lint": "fecs src/ --type=js"
   },
   "devDependencies": {
-    "babel-core": "^6.7.6",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-transform-runtime": "^6.7.5",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-core": "^6.24.0",
+    "babel-loader": "^6.4.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-es2015": "^6.24.0",
+    "babel-preset-stage-2": "^6.22.0",
     "babel-runtime": "^5.8.38",
     "css-loader": "^0.16.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.5",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "file-loader": "^0.10.1",
     "html-webpack-plugin": "^2.17.0",
     "less": "^2.6.1",
     "less-loader": "^2.2.3",
-    "style-loader": "^0.13.1",
-    "vue-hot-reload-api": "^1.3.2",
-    "vue-html-loader": "^1.2.2",
-    "vue-loader": "^8.2.2",
-    "vue-style-loader": "^1.0.0",
-    "webpack": "^1.12.15"
+    "style-loader": "^0.14.1",
+    "vue-loader": "^11.1.4",
+    "vue-style-loader": "^2.0.4",
+    "vue-template-compiler": "^2.2.4",
+    "webpack": "^2.2.1",
+    "webpack-dev-server": "^2.4.2"
   },
   "dependencies": {
-    "vue": "^1.0.21",
-    "vue-router": "^0.7.13",
-    "vue-resource": "^0.7.0"
+    "vue": "^2.2.4",
+    "vue-resource": "^1.2.1",
+    "vue-router": "^2.3.0",
+    "vuex": "^2.2.1"
+  },
+  "babel": {
+    "presets": [
+      "es2015",
+      "stage-2"
+    ],
+    "plugins": [
+      "transform-runtime"
+    ]
   }
 }
