@@ -1,55 +1,69 @@
 <template>
-
 <form action="" class="setting">
     <p class="form-control">
-        <label for="">页数</label>
-        <input type="text" name="perPage" value="5" class="input" v-model="perPage">
+        <label for="perPage">页数</label>
+        <input type="text" id="perPage" name="perPage" class="input" v-model="config.perPage" @input="fakeUpdateConfig" />
     </p>
 
     <p class="form-control">
-        <label for="">更新天数</label>
-        <input type="text" name="expired" value="1" class="input" v-model="expired">
+        <label for="expired">更新天数</label>
+        <input type="text" id="expired" name="expired" value="1" class="input" v-model="config.expired" @input="fakeUpdateConfig" />
     </p>
 
-    <p class="form-control"><button type="submit" class="button save" v-on:click="save($event)">保存</button></p>
+    <p class="form-control">
+        <button type="submit" class="button save" v-on:click="save">保存</button>
+    </p>
 </form>
-
 </template>
 
 <script>
-    var localStorage = require('../util/localstorage');
+import { mapGetters, mapActions } from 'vuex'
 
-    module.exports = {
-        data: function () {
-            return {
-                perPage: 5,
-                expired: 1
-            }
+export default {
+    computed: {
+        ...mapGetters([
+            'config'
+        ])
+    },
+
+    methods: {
+        ...mapActions([
+            'updateConfig',
+        ]),
+
+        fakeUpdateConfig(e) {
+            const cfg = {}
+
+            cfg[e.target.id] = e.target.value
+
+            this.updateConfig(cfg, false)
         },
 
-        methods: {
-            save: function (e) {
-                e.preventDefault();
+        save(e) {
+            e.preventDefault()
 
-                if (this.pageNum <= 0) {
-                    alert('pageNum must > 0');
-                    return;
-                }
-
-                if (this.expired < 0)  {
-                    alert('更新天数 >= 0');
-                    return;
-                }
-
-                localStorage.set('options', {
-                    perPage: this.perPage,
-                    expired: this.expired
-                });
-
-                alert('设置成功！');
+            if (this.perPage <= 0) {
+                alert('perPage must > 0')
             }
+            else {
+                this.updateConfig({
+                    perPage: this.config.perPage
+                })
+            }
+
+            if (this.expired <= 0)  {
+                alert('更新天数 > 0')
+            }
+            else {
+                this.updateConfig({
+                    expired: this.config.expired
+                })
+            }
+
+            alert('设置成功')
         }
-    };
+    }
+}
 </script>
 
 <style lang="less">
